Fix alignment for NetShareEnum request handling
The missing alignment meant that the info level was being misread if the host
name of the server had an even number of characters, because in that case the
end of the string ended up not being 4-byte-aligned.

This meant that the info level was being misread as e.g. 65536 instead of 1,
which had the ultimate result that instead of success and a list of the
available shares, an error message would be sent back and the server couldn't be
accessed.