Fix off-by-one error in SearchContext.restartAt(FileInfo) code
The problem is that in order to do the comparison between the file name from the
iterator and the passed-in FileInfo, we always need to consume the respective
Path from the DirectoryStream iterator, so even when the comparison declares
success, we have already consumed that Path. This means that the next FIND_NEXT2
call will skip that entry and start one Path too late when it accesses this
search context again.

For correctness, we therefore have to rewind the iterator by one after having
found the correct Path, so that the next nextFileInfo() call restarts at the
right point.

The current solution isn't the best for large directories performance-wise due
to having to iterate twice over all files up to the resume point, but at least
it is correct and better than omitting the first file from every FIND_NEXT2
response (or equivalent split-up Find request reply for SMB2).